msm_performance: disable touchboost by default
This is crazy overkill (boosts to either 1344MHz or 1555MHz for 1.5s
or something like that) so not great for battery but users can turn
it back on if they don't want the loss of performance.

Also, make sure that the max frequency can't be tampered with when
touchboost is disabled.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>